requirements.txt
facenet-pytorch>=2.5.2
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=8.0.0
matplotlib>=3.4.0
opencv-python>=4.5.0
scipy>=1.7.0 